Adjustments after review
